changes default ports, refactors bit masking to use int instead of signed bytes, adds readme and couple of assets

+++ src/main/java/tcl/itm/ITMBase.java
@@ -4,10 +4,10 @@ package tcl.itm;
  * for ITM Trace packet format see  http://infocenter.arm.com/help/index.jsp?topic=/com.arm.doc.ddi0314h/Chdbicbg.html
  */
 public interface ITMBase {
-    byte header_size_bitmask = 0x3;
-    byte header_mandatory_field_bitmask = 0x4;
-    byte header_mandatory_field_value = 0x0;
-    byte header_source_address_bitmask = (byte) 0xF8;
-    byte header_source_address_shift_offset = 3;
+    int header_size_bitmask = 0x3;
+    int header_mandatory_field_bitmask = 0x4;
+    int header_mandatory_field_value = 0x0;
+    int header_source_address_bitmask =  0xF8;
+    int header_source_address_shift_offset = 3;
 
 }
